Replace RecipeCard with StyledRecipeCard for improved styling and tap functionality
